backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Supabase
supabase==2.3.4
postgrest==0.13.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Vector Database & Embeddings
pgvector==0.2.4
openai==1.10.0
anthropic==0.18.1
sentence-transformers==2.3.1

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
pillow>=11.0.0
pytesseract==0.3.10
pdf2image==1.17.0

# OCR Services (optional)
google-cloud-vision==3.6.0
boto3==1.34.34  # For AWS Textract

# Audio Processing
openai-whisper==20231117
pydub==0.25.1

# Task Queue
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pyjwt==2.8.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.39.2
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
